fix(1369): rename folder
