+++ frontend/src/components/Boards/MyBoards/ListPersonalBoards/EmptyPersonalBoards/index.tsx
